Rename the indirect host counting test file
